pyproject.toml
[project]
name = "djangosige"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-crispy-forms>=2.4",
    "geraldo",
    "pysignfe",
    "python-decouple>=3.8",
    "signxml>=4.2.0",
]

[tool.uv.sources]
geraldo = { url = "https://github.com/thiagopena/geraldo/archive/master.zip" }
pysignfe = { url = "https://github.com/thiagopena/PySIGNFe/archive/master.zip" }

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "factory-boy>=3.3.3",
    "pre-commit>=4.3.0",
    "ruff>=0.14.3",
]

[tool.coverage.run]
include = ["djangosige/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]


[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"
